Extract useResizeObserver hook
