+++ packages/react/use-debounce-callback/README.md
@@ -1,0 +1,13 @@
+# `react-use-debounce-callback`
+
+## Installation
+
+```sh
+$ yarn add @radix-ui/react-use-debounce-callback
+# or
+$ npm install @radix-ui/react-use-debounce-callback
+```
+
+## Usage
+
+This is an internal utility, not intended for public usage.
