+++ packages/react/use-resize-observer/README.md
@@ -1,0 +1,13 @@
+# `react-use-resize-observer`
+
+## Installation
+
+```sh
+$ yarn add @radix-ui/react-use-resize-observer
+# or
+$ npm install @radix-ui/react-use-resize-observer
+```
+
+## Usage
+
+This is an internal utility, not intended for public usage.
