+++ packages/react/use-resize-observer/src/index.ts
@@ -1,0 +1,1 @@
+export { useResizeObserver } from './useResizeObserver';
